added counting code brevity

+++ src/main/java/ru/dexterity/dao/models/TaskRating.java
@@ -30,9 +30,9 @@ public class TaskRating {
 
     private String solution;
 
-    private Double brevity;
-    private Double rapidity;
-    private Double resourceConsumption;
-    private Double totalScore;
+    private int brevity;
+    private double rapidity;
+    private double resourceConsumption;
+    private double totalScore;
 
 }
